Create basic view for game
Simple html view with a title and a single button to initiate gameplay. No forms/input fields yet; the initial iteration of the game will simply be run through prompts and alerts.

The app.js script is linked in the head section, but contains no functionality to interact with the DOM yet.