Wire up gameplay functionality
Set listeners in app.js: when DOM content has loaded, set 'click' listener on the 'play-round' button. Click triggers prompts for the player's bet and their guess - prompts ask for number within specific intervals, but not validations are provided. After receiving the player's input, the .playRound() method of bettingGame is called, which updates the play'ers bank balance (playerBankroll) and tells them the results of their guess with an alert message.