Improved HTML view with basic form
The second iteration of the game will allow the user to play by inputting their guess and their bet to form fields on the page instead of responding to prompts. This commit adds a basic form:
- a dropdown menu to take bet amounts
- an input field to take a guess between 1 and 10
- the 'Play Round' button is preserved from the previous version, but put inside a div to facilitate future styling
- a 'Restart Game' button

app.js does not yet reference form fields, but the game will still work in its previous form via the 'Play Round' button.