Change bettingGame (object) into constructor function that can be called to instantiate new games
Previously, there was no logic present to allow the player to reset the game and start again. Converting the initial bettingGame object into a constructor function, which is called once when the page loads, will allow for new games to start.

The new constructor function also takes an argument, which is assigned to be the player's starting balance variable. This will allow for future games to be started with different initial amounts.